Add files via upload
Analyzing the monthly trend of accident, there has been a gradual increase in the number of accidents between 2021- 2022. However, there was Major, Minor, Severe and Moderate.
Donut Chart was used to determine the percentage of the accident in five cities and three state in US, an insight was determined to know the state that have high concentrations of Accident.